+++ spec/fdp_index_proxy_spec.rb
@@ -1,12 +1,7 @@
-# # frozen_string_literal: true
-# require 'rspec'
+# frozen_string_literal: true
 
-# RSpec.describe FdpIndexProxy do
-#   it "has a version number" do
-#     expect(FdpIndexProxy::VERSION).not_to be nil
-#   end
-
-#   it "does something useful" do
-#     expect(false).to eq(true)
-#   end
-# end
+RSpec.describe FdpIndexProxy do
+  it "has a version number" do
+    expect(FdpIndexProxy::VERSION).not_to be_nil
+  end
+end